# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp Order Management API for Nairobi SMEs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders with optional filters
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: customerId
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Manually create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order with items
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
...
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate Mpesa STK push for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiatePaymentRequest"
      responses:
        "200":
          description: STK push initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitiateResponse"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Payment record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found

  # ── Analytics ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]

    # ── Enums ─────────────────────────────────────────────────────────────
    OrderStatus:
      type: string
      enum: [pending, confirmed, paid, preparing, ready, delivered, cancelled]
    PaymentStatus:
      type: string
      enum: [pending, processing, completed, failed, cancelled, timeout]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        status:
          $ref: "#/components/schemas/OrderStatus"
        totalAmount:
          type: number
        currency:
          type: string
          default: KES
        rawMessage:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        assignedToId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, customerId, status, totalAmount, currency, createdAt, updatedAt]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        variantId:
          type: integer
          nullable: true
        productName:
          type: string
        variantName:
          type: string
          nullable: true
        quantity:
          type: number
        unitPrice:
          type: number
        totalPrice:
          type: number
      required: [id, orderId, productId, productName, quantity, unitPrice, totalPrice]
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
            customer:
              $ref: "#/components/schemas/Customer"
            payment:
              $ref: "#/components/schemas/Payment"
              nullable: true
          required: [items, customer]
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [orders, meta]
...
    CreateOrderRequest:
      type: object
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              productId: { type: integer }
              variantId: { type: integer, nullable: true }
              quantity: { type: number }
            required: [productId, quantity]
        notes:
          type: string
      required: [customerId, items]
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
        notes:
          type: string
      required: [status]

    # ── Products ──────────────────────────────────────────────────────────
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        whatsappPhone:
          type: string
        name:
          type: string
          nullable: true
        totalOrders:
          type: integer
        totalSpend:
          type: number
        loyaltyPoints:
          type: integer
        lastOrderAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, whatsappPhone, totalOrders, totalSpend, loyaltyPoints, createdAt]
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        amount:
          type: number
        currency:
          type: string
          default: KES
        status:
          $ref: "#/components/schemas/PaymentStatus"
        mpesaPhone:
          type: string
          nullable: true
        merchantRequestId:
          type: string
          nullable: true
        checkoutRequestId:
          type: string
          nullable: true
        mpesaReceiptNumber:
          type: string
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, orderId, amount, currency, status, createdAt]
    InitiatePaymentRequest:
      type: object
      properties:
        orderId:
          type: integer
        phone:
          type: string
          description: Mpesa phone number (07XX or 254XX format)
      required: [orderId, phone]
    PaymentInitiateResponse:
      type: object
      properties:
        paymentId:
          type: integer
        checkoutRequestId:
          type: string
        message:
          type: string
      required: [paymentId, checkoutRequestId, message]

    # ── Analytics ─────────────────────────────────────────────────────────
...