# lib/api-zod/src/generated/types/initiatePaymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Order Management API for Nairobi SMEs
 * OpenAPI spec version: 0.1.0
 */

export interface InitiatePaymentRequest {
  orderId: number;
  /** Mpesa phone number (07XX or 254XX format) */
  phone: string;
}